Rename(*): Change locs and names
