+++ src/components/Footer/Footer.js
@@ -1,9 +1,9 @@
-import "../styles/footer.css";
-import "../styles/reset.css";
-import twitterIcon from "../assets/akar-icons_twitter-fill.svg";
-import facebookIcon from "../assets/akar-icons_facebook-fill.svg";
-import youtubeIcon from "../assets/akar-icons_youtube-fill.svg";
-import instagramIcon from "../assets/akar-icons_instagram-filled.svg";
+import "./Footer.style.css";
+import "../../styles/reset.css";
+import twitterIcon from "../../assets/akar-icons_twitter-fill.svg";
+import facebookIcon from "../../assets/akar-icons_facebook-fill.svg";
+import youtubeIcon from "../../assets/akar-icons_youtube-fill.svg";
+import instagramIcon from "../../assets/akar-icons_instagram-filled.svg";
 
 function Footer() {
   return (
