+++ src/components/SearchBar/SearchBar.style.css
